build

+++ src/main/java/com/ecommerce/repository/StateRepository.java
@@ -11,7 +11,7 @@ import com.ecommerce.entity.State;
 
 @CrossOrigin("http://localhost:4200")
 @RepositoryRestResource
-public interface StateRepository extends JpaRepository<State, Integer>{
+public interface StateRepository extends JpaRepository<State, Long>{
 
     List<State> findByCountryCode(@Param("code") String code);
